Reduce the size of the Docker build
This corrects several mistakes in the original Docker implementation, and
significantly reduces the resulting image size. It also builds significantly
faster.

We also upgrade some of the requirements to more recent versions that have
binary packages available, so that we don't have to install the full compiler
suite within the container.